Restructure build-system to only publish important/usable modules
There isn't really much a point to publishing either template or compat submodules to the repository server as they shouldn't be depended on anyway, I think we should also consider not publishing some of the platform impls as well, fabric/velocity, for example, can't and shouldn't be used at all. (Not sure about bukkit). I decided to come up with this because we now have platforms which are external (sponge/bungee) and changed artifacts, making it not really possible to publish them at all.

Not sure if extensions#publishJavaComponents should be kept or not, can be removed if you want.

If we agree on something here, I'll open a PR to ViaBackwards as well and do the same for ViaRewind.