Handle attribute_modifiers inside container components in 1.21->1.20.5
Since items inside the container component can also contain attribute modifiers, we need to handle them as well since the client would get kicked otherwise for the missing id nbt. While this works for now, we should rewrite our 1.21+ component rewriters to fully parse the item as soon as the ComponentRewriter1_20_5 is finished.

Closes https://github.com/ViaVersion/ViaVersion/issues/3968