Fix ProtocolVersion#isKnown javadoc to make difference with isRegistered more clear
The current javadoc is wrong as it indicates that ProtocolVersion#getProtocol will return a protocol instance with -1 as protocol id, which is never the case. From what I see the function is used as check if a protocol version instance is equal to the unknown protocol version, and should therefore be explained in the javadoc.