Properly track minecart object entity types in <= 1.12.2
Fixes furnace and command block minecart entity data not being translated in 1.8->1.9 as the entity data index checks for the correct type, but we didn't take in account the data field for tracking object types. This also removes the getTypeFromId() function and replaces its usages with the sub functions from EntityType/ObjectType.