Make warning when no working pipe is found more clear.
We had this in the past on our Discord server where people didn't understand what the original log message means. Since this is only relevant for serverside platforms (as clientside impls usually just have all jars inbuilt) I decided to simplify this message and to explicit mention ViaBackwards/ViaRewind as those are the only serverside supported Via addons we had in the past years (except ViaAprilFools but shrug).