Drop (block) entity data in 1.20.3->.5 if no id is present
Would appreciate if you could also verify that, but from looking at the game code I'm quite sure that the server/client need a valid id for the data to be read correctly. Meaning it's wrong on our side to try to append the id (as won't work in all cases anyway) and rather just drop data without id (as then unused on the client anyway).