Cancel unknown block entities in BlockRewriter#registerBlockEntityData
Mojang changed block entity type parsing in 1.20.5 to throw on invalid block entities instead of ignoring them, we forgot to update our handle for newer protocols which will now kick if you join with a 1.21.4 client on a 1.21.5 server and set a test block. As this change should not hurt in older versions (as an invalid id would also cause the data to be ignored) I decided to put this directly instead of having multiple register methods.